pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-shared"
version = "1.0.0"
description = "Shared utilities for RAG orchestration services"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "QIG", email = "dev@qig.ai" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core logging (no external deps needed for basic logging)

    # Redis/Dragonfly client
    "redis>=5.0.0",
]

[project.optional-dependencies]
# OpenTelemetry observability stack
observability = [
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.37.0",
    "opentelemetry-instrumentation>=0.58b0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-requests>=0.58b0",
    "opentelemetry-instrumentation-httpx>=0.58b0",
    "opentelemetry-instrumentation-logging>=0.58b0",
    "opentelemetry-semantic-conventions>=0.58b0",
]

# Flask support (if needed)
flask = [
    "opentelemetry-instrumentation-flask>=0.58b0",
]

# Database utilities
database = [
    "asyncpg>=0.29.0",
]

# All extras combined
all = [
    "rag-shared[observability,flask,database]",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourorg/rag-shared"
Repository = "https://github.com/yourorg/rag-shared"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_shared"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
